Adjust location of buttons
I made them fixed position, so if a large amount of content is pasted, they will always be shown even if the user scrolls down a lot. The Format button is now toward the right side of the text box, so it should be much less likely for it to be on top of text. Similarly, I gave more space for the "instructions" button and moved it to the right.